Change the structure of car details
